requirements.txt
python-decouple==3.1
gunicorn==19.9.0
Flask==1.0.2
Flask-Cors==3.0.7
psycopg2==2.8.2
SQLAlchemy==1.3.3
GeoAlchemy2==0.6.1
geomet==0.2.0.post2
